backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6

# Core dependencies
spotipy>=2.24.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.1.0

# Additional utilities
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
packaging>=23.2

